feat(native_render): rename native render
