+++ renderer/native/android/src/main/jni/include/render/native_render_manager.h
@@ -1,55 +1,96 @@
-/*
- *
- * Tencent is pleased to support the open source community by making
- * Hippy available.
- *
- * Copyright (C) 2019 THL A29 Limited, a Tencent company.
- * All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
-
 #pragma once
 
-#include <jni.h>
+#include <memory>
 
-class NativeRenderManager {
+#include "jni/scoped_java_ref.h"
+
+#include "dom/dom_node.h"
+#include "base/macros.h"
+#include "dom/render_manager.h"
+#include "dom/serializer.h"
+
+namespace hippy {
+inline namespace dom {
+
+class NativeRenderManager : public RenderManager {
  public:
-  static void Init();
+  NativeRenderManager(std::shared_ptr<JavaRef> render_delegate);
 
-  static void Destroy();
+  virtual ~NativeRenderManager() = default;
+  NativeRenderManager(const NativeRenderManager &) = delete;
+  NativeRenderManager &operator=(const NativeRenderManager &) = delete;
+  NativeRenderManager(NativeRenderManager &&) = delete;
+  NativeRenderManager &operator=(NativeRenderManager &&) = delete;
+
+  int32_t GetId() { return id_; }
+
+  void CreateRenderNode(std::vector<std::shared_ptr<DomNode>>&& nodes) override;
+  void UpdateRenderNode(std::vector<std::shared_ptr<DomNode>>&& nodes) override;
+  void DeleteRenderNode(std::vector<std::shared_ptr<DomNode>>&& nodes) override;
+  void MoveRenderNode(std::vector<std::shared_ptr<DomNode>>&& nodes) override;
+  void UpdateLayout(const std::vector<std::shared_ptr<DomNode>>& nodes) override;
+  void MoveRenderNode(std::vector<int32_t>&& moved_ids, int32_t from_pid, int32_t to_pid) override;
+  void EndBatch() override;
+
+  void BeforeLayout() override;
+  void AfterLayout() override;
+
+  using DomValue = tdf::base::DomValue;
+
+  void AddEventListener(std::weak_ptr<DomNode> dom_node, const std::string& name) override;
+
+  void RemoveEventListener(std::weak_ptr<DomNode> dom_node, const std::string& name) override;
+
+  void CallFunction(std::weak_ptr<DomNode> dom_node, const std::string& name, const DomArgument& param,
+                    uint32_t cb_id) override;
+
+  void SetDensity(float density) { density_ = density; }
+  float GetDensity() { return density_; }
+
+  void SetDomManager(std::weak_ptr<DomManager> dom_manager) { dom_manager_ = dom_manager; }
+  std::shared_ptr<DomManager> GetDomManager() const { return dom_manager_.lock(); }
+
+  static void Insert(const std::shared_ptr<NativeRenderManager>& render_manager);
+  static std::shared_ptr<NativeRenderManager> Find(int32_t id);
+  static bool Erase(int32_t id);
+  static bool Erase(const std::shared_ptr<NativeRenderManager>& render_manager);
+
+ private:
+  inline void MarkTextDirty(uint32_t node_id);
+
+  inline float DpToPx(float dp) const;
+
+  inline float PxToDp(float px) const;
+
+  void CallNativeMethod(const std::pair<uint8_t*, size_t>& buffer, const std::string& method);
+
+  void CallNativeMethod(const std::string& method);
+
+  void CallNativeMeasureMethod(const int32_t id, const float width, const int32_t width_mode, const float height,
+                               const int32_t height_mode, int64_t& result);
+
+  struct ListenerOp {
+    bool add;
+    std::weak_ptr<DomNode> dom_node;
+    std::string name;
+
+    ListenerOp(bool add, std::weak_ptr<DomNode> dom_node, const std::string& name) {
+      this->add = add;
+      this->dom_node = dom_node;
+      this->name = name;
+    }
+  };
+
+  void HandleListenerOps(std::vector<ListenerOp>& ops, const std::string& method_name);
+
+ private:
+  int32_t id_;
+  std::shared_ptr<JavaRef> render_delegate_;
+  std::shared_ptr<tdf::base::Serializer> serializer_;
+  float density_ = 1.0f;
+  std::vector<ListenerOp> event_listener_ops_;
+
+  std::weak_ptr<DomManager> dom_manager_;
 };
-
-jint OnCreateNativeRenderProvider(JNIEnv* j_env,
-                                  jobject j_obj,
-                                  jfloat j_density);
-
-void OnDestroyNativeRenderProvider(JNIEnv* j_env,
-                                   jobject j_object,
-                                   jint j_instance_id);
-
-void UpdateRootSize(JNIEnv* j_env, jobject j_obj, jint j_instance_id,
-                    jfloat width, jfloat height);
-
-void UpdateNodeSize(JNIEnv* j_env, jobject j_obj, jint j_instance_id, jint j_node_id,
-                    jfloat width, jfloat height, jboolean j_is_sync);
-
-void OnReceivedEvent(JNIEnv *j_env, jobject j_object,
-                     jint j_instance_id, jint j_root_id, jint j_dom_id, jstring j_event_name,
-                     jbyteArray j_buffer, jint j_offset, jint j_length,
-                     jboolean j_use_capture, jboolean j_use_bubble);
-
-void DoCallBack(JNIEnv *j_env, jobject j_object,
-                jint j_instance_id, jint j_result, jstring j_func_name, jint j_node_id,
-                jlong j_cb_id, jbyteArray j_buffer, jint j_offset, jint j_length);
+}  // namespace dom
+}  // namespace hippy
